Add new CStore indomaret response model extending BaseResponse
