+++ corekit/src/main/java/com/midtrans/sdk/corekit/core/api/snap/model/pay/response/BasePaymentResponse.java
@@ -140,6 +140,16 @@ public class BasePaymentResponse implements Serializable {
     @SerializedName("gopay_expiration_raw")
     protected String gopayExpirationRaw;
 
+    /**
+     * Indomaret
+     */
+    @SerializedName("payment_code")
+    protected String paymentCode;
+    @SerializedName("store")
+    protected String store;
+    @SerializedName("indomaret_expire_time")
+    protected String indomaret_expire_time;
+
     public ArrayList<String> getValidationMessages() {
         return validationMessages;
     }
@@ -151,6 +161,5 @@ public class BasePaymentResponse implements Serializable {
     public String getStatusMessage() {
         return statusMessage;
     }
-
 
 }
